Trail 2

+++ cfm/cfm-code/FragmentGraph.h
@@ -334,9 +334,9 @@ protected:
     void getSampledTransitionIdsDifferenceWeightedBFS(std::set<unsigned int> &selected_weights, std::set<int> &visited,
                                                           int frag_id, std::vector<int> &path, std::set<int> &selected_ids);
 
-    void getCommonAncestors(std::set<double> &selected_weights, std::set<int> &visited,
-                                        int frag_id, std::vector<std::pair<int, int>> &path,
-                                        std::map<int, std::vector<int>> &trans_to_interest_frags_map);
+    void getCommonAncestors(std::set<unsigned int> &selected_weights, std::set<int> &visited,
+                            int frag_id, std::vector<std::pair<int, int>> &path,
+                            std::map<int, std::vector<int>> &trans_to_interest_frags_map);
 
     bool is_match(std::set<unsigned int> &weights, double mass) const;
